fix: add method interceptor to manage child node removal in UInjector
React removes child nodes before unmounting the root element a component renders. This caused React in trying to remove deeply nested child nodes from the wrapping container before running our cleanup function which moves nodes back to the wrapping container.
An interceptor on the `removeChild` method allows us to test whether we've manipulated with the child in question and "remove" it from our change list and removing it from the current host, which is the injection target.