fix: switch to type-only exports in Actions index
Issue: The value exports caused tests (and potentially other code) to fail, because only typings were exported and the implementation couldn't be found. Type-only exports provides additional context to typescript to not look up for any implementation, and instead, only allow usage as a type, rather than as a value.